nested set refactoring
NestedSet class is used to group nodes. a set must have a node which is
the root of the represented tree.Hence for each set a node is created.
To enforce integrity : 1 - a set must have a node (to avoid set with no
node , which means group with no nodes , a group exist because a nodes)
2 - a node must have set
1 & 2 create a reflexive dependency which is impossible to satisfy.To
resolve this , we will use the set(group) identifier only in the node.
To follow exactly the nested set model , we will rename NestedSetNode to
NestedSet